add register_credential.proto, re-run proto generation
here we add register_credential.proto and update proto/gen_attest.sh

The other diffs are generated code.

The CertifiedBlob message is the proof from solving the EK-based TPM registration
challenge detailed in
https://trustedcomputinggroup.org/wp-content/uploads/EK-Based-Key-Attestation-with-TPM-Firmware-Version-V1-RC1_9July2025.pdf#page=8.

The SevSnpSvsmAttestation message contains all the information to prove
the binding between the hardware SNP attestation report and the TPM. See
https://www.amd.com/en/developer/sev.html to find a link to the SVSM
spec for the format of the vtpm_service_manifest.