implement EK-based Key attestation for the SVSM e-vTPM
The client server files for import_certify.go corresponds to the spec
for EK-based Key attestation with firmware version.
See https://trustedcomputinggroup.org/wp-content/uploads/EK-Based-Key-Attestation-with-TPM-Firmware-Version-V1-RC1_9July2025.pdf#page=8

This is to be used with Coconut SVSM's single service attestation for
the vTPM. We use SVSM's service attestation and the SNP hardware
attestation report to create trust for the EK, instead of an EK
certificate.
See https://github.com/coconut-svsm/svsm