modify attest and verify CLI for SVSM
We add the --svsm flag to `gotpm attest` which attaches the SVSM tee
attestation onto the Attestation proto.

For verify we add the --skip-gce-technology flag to skip verifying the
SNP attestation report in the gotpm verify debug command. Instead, we
expect users to use either go-sev-guest or gce-tcb-verifier to verify
the SNP attestation report.