use proto helpers for read/write msgs in attest and verify cli
We also break the use of the format flag for modifying --input in
the verify debug cmd. This is because the format flag should only be
used for modifying output format per its definition in flags.go.

We also do some minor code improvement with adding reportNonce to avoid
duplicating code for initialzing the tee attestation options.

No other functional change intended.